requirements.txt
langchain>=0.0.300
aiohttp>=3.8.0
pandas>=1.3.0
tqdm>=4.65.0
python-dotenv>=0.19.0